refactored out common interface for session/course lms.
